config: 카카오 로그인 sdk 추가
