+++ test/widget_test.dart
@@ -7,13 +7,12 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-
-import 'package:three_days_flutter/main.dart';
+import 'package:three_days/three_days_app.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const MyApp());
+    await tester.pumpWidget(const ThreeDaysApp());
 
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);
